requirements.txt
absl-py==1.2.0
actionlib==1.13.2
angles==1.9.13
astunparse==1.6.3
berserk==0.10.0
bondpy==1.8.6
bprofile==1.3.4
cachetools==5.2.0
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2022.6.15
charset-normalizer==2.1.0
chess==1.9.2
chess-board==0.3.1
click==8.1.3
click-log==0.4.0
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.16.0
cvzone==1.3.3
cycler==0.11.0
Deprecated==1.2.13
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.2
flatbuffers==1.12
fonttools==4.34.4
future==0.18.2
gast==0.4.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
image-geometry==1.16.0
imageio==2.20.0
imgaug==0.4.0
importlib-metadata==4.12.0
imutils==0.5.4
interactive-markers==1.12.0
iso8601==1.0.2
joblib==1.1.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
laser_geometry==1.6.7
libclang==14.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
message-filters==1.15.14
ndjson==0.3.1
networkx==2.8.5
numpy==1.23.1
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.4
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-qt-binding==0.4.4
pyttsx3==2.90
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.28.1
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.14
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
sensor-msgs==1.13.1
Shapely==1.8.2
six==1.16.0
sklearn==0.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
stockfish==3.28.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.1.0
tifffile==2022.5.4
topic-tools==1.15.14
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.0
wrapt==1.14.1
xacro==1.14.13
zipp==3.8.1
